source/av: handle multiple audio streams w/ proper sample rate
We need to tell sox to use and resample to the _player_ format
instead of the source format.  Otherwise 48000 Hz audio (common for
video?) sounds very slow when attempting to play back on the default
44100 Hz.  It is also likely preferable to choose the audio stream
which matches the player channel count instead of letting sox
automatically invoke the remix effect.

Before this change, playing 6-channel, 48000 Hz source into a sink
expecting stereo, 44100 Hz is rather disturbing...

While we're at it, clarify the spawn code for the sox source, too.

We can drop the test_format_from_file test now, since
it's probably overkill at this point.