source/sox: cleanup error handling, quiet redundant warnings
Showing the warning for the same file over and over again is
annoying when seeking a file handled by avconv/ffmpeg, so stop
doing it.

While we're at it, the error handling for __load_comments is totally
redundant (absent of race conditions if another process modifies the
file).  However, if we do hit races, all the other soxi invocations
would be racy, too; but we really just shouldn't care about such
a corner case in here.