rework packaging to use GNU make + gemspec instead of Hoe
Currently, this allows us to use different manpage paths for the
tarball and gem; as gem-man and setup.rb expect different paths
for manpages.

Additionally, Hoe is designed for Ruby projects.  dtas may include
Perl/shell/Python/whatever in the future.  So use GNU make as it
is more suited for language agnosticism.