player: support seeking based on embedded cuesheet (FLAC)
This adds the ability to seek internally within FLAC file
based on the internal CUE sheet.  Other formats may be supported
in the future, but FLAC is the only one I know of which supports
embedded cue sheets.

Note: flac 1.3.0 is recommended for users of non-CDDA-compatible
formats.

See updates to dtas-player_protocol(7) for details.