buffer: remove ioctl syscall in common paths
We do not need this for single sink situations (the common case)
at all.  We also do not need to check IO#nread for splice, either;
we can just do non-blocking I/O.  The only common path where we
might still need it is the non-splice case with multiple sinks.