dtas-partstats: initial implementation
dtas-partstats divides large audio files into small partitions (10
seconds by default) and runs the "stats" effect of sox(1) against each
partition.

Currently it emits space-delimited output, but configurable output
options (including Sequel/SQLite) support is coming.

The intended use of this tool is for quickly finding the loudest
portions of a given recording without the need for a graphical viewer.
This can be useful for selectively applying (and testing the results of)
dynamic range compression filters.

Use with sort(1) in a pipeline is recommended in this scenario
(but again, Sequel support is coming).