merge some select calls for splice users
Avoid an additional select syscall in the splice path by injecting
the target checks into the main loop.  We can do this because we
always process writers before readers.  This adds additional
userspace processing, but it avoids one potentially expensive (and
potentially task-switching) syscall in many cases; so it should be
worth it to avoid skipping with small buffer.

This should avoid buffer underuns with mixed-sized buffers when
using multiple sinks.

This could work for the read_write path, too, but I don't
use that enough and this change may still be buggy even
for splice users.