use omap (ordered map) for env hashes
While the Ruby Hash class is ordered in 1.9+, the YAML
specifications do not specify hashes as ordered by default.
Thus we must explicitly declare ordering via !omap for
interopability with non-Ruby tools.

This makes the YAML output of dtas-sourcedit and dtas-sinkedit
slightly more verbose

Users of dtas-splitfx are also encouraged to declare !omap
when creating their YAML files for interoperability.

Ordering env is important because any implementation of
built-in variable expansion is dependent on it.