player: implement software volume control
This is on a linear scale from 0.0 (mute) to 1.0 (no change)
This is in the MPRIS spec and mpd as well (scaled to 0-100)

This changes dtas-console key bindings (0/9) slightly to match
mplayer more closely.  ReplayGain preamp adjustment has moved from
'0'/'9' to '7'/'8' keys.  The 'm' key also toggles mute state (the
pre-mute volume is stored in the dtas-console instance, not in
dtas-player itself).