tracklist: reshuffle tracklist when exhausted
For now, shuffle means "random without repeats".  So avoid repeating
ourselves by shuffling when we've reached the end of the shuffled
tracklist and have no more tracks to go.  Whenever a client restarts
the player or hits the loop for repeat, we'll have a freshly
shuffled list for them to listen to.