dtas-tl: hopefully fix up encoding issues
POSIX filesystems allow any combination of bytes, so we should be
able to edit the tracklist when we have a filename with non-UTF-8
(or whatever locale the user uses).  Try to present the user with
a reasonable name when they have an external encoding (typically
UTF-8); but be prepared to dump out whatever binary sequence
the filesystem allows.