dtas 0.13.1 - dtas-tl encoding bugfixes
This release addresses encoding and escaping-related bugs affecting
the use of "dtas-tl edit" (and to a lesser extent, "dtas-tl cat").
Filenames with characters requiring shell-escaping were not
round-tripped correctly, causing movement/addition of such files
(which may not be properly encoded to any encoding) to be handled
incorrectly.

There are also some mlib bug fixes, but the mlib API and CLI are
currently unstable and undocumented...

4 changes since 0.13.0:
      dtas-tl: hopefully fix up encoding issues
      dtas-tl: shell-unescape cat output
      mlib: fix foreign-key ordering issues with delete
      dtas-mlib: avoid redundant '/' in dumped pathname