dtas 0.14.0 - tracklist fixes and improvements
The player tracklist now has a "consume" mode similar to
what exists mpd.  The advantage of using "tl consume"
over the plain queue is the tracklist is more easily
editable.  You may enable or disable it in a running
player using:

	dtas-tl consume <true|false>

This also fixes an output formatting bug in "dtas-tl cat"
introduced in 0.13.1.

Documentation is converted to Perl POD documentation format.
It should be easier to build the documentation as pod2man
is nearly universal nowadays, unlike pandoc.

14 changes since 0.13.1:

      doc: update-footer.rb: fix for frozen_string_literal
      doc: convert to perlpod(1) from Markdown
      doc: preserve times in website
      dtas-tl: fix "cat" output
      player: avoid allocation on sort
      String#split with a single-byte instead of regexp
      source/sox: explicitly freeze comment keys
      source/sox: try_to_fail_harder is a normal method
      mlib: support all formats of player
      mlib: support --force option
      build: install-gem forced to "--local" domain
      mlib: use more appropriate methods
      player: support "consume" mode for tracklist
      doc: update documentation for tl consume and shuffle

Many more releases on the horizon!