rg_state: fix no-op gain detection
Seeing tiny (or zero) values for the "gain" effect is an eyesore
in process table output, and a waste of CPU cycles in sox.  So
stop using the "gain" effect for imperceptible changes in volume.

While we're at it, remove the pointless knobs for setting this,
too.  They were never documented and I doubt anybody would want
to tweak them at runtime.