TODO: add a note about using a more stable glue language
Every Ruby release is a constant treadmill of minor
incompatibilities which require busywork and maintenance every
year.

It's been roughly a decade after the major 1.8 -> 1.9 breakages
which I was naive enough to believe was one time thing.  Now
with Ruby 3.0 incompatibilities on the horizon, I've had enough.

Perl5 is the most logical language since most of my other
projects are Perl5, and we already have dtas-graph implemented
in it.  But, implementing my own glue language for this isn't
out of the question, either...