doc: drop ordered map from examples
"omap" is specific to Ruby and makes interopability with other
languages more difficult.

While it's true environment variables are stored as an ordered
array of C strings (see environ(7)); order doesn't matter in
practice.  Everyone in the real world treats the environment as
an unordered key-value store, and we shall follow.