player: reduce syscalls when splicing to single target
splice(2) alone does not give enough information as to whether
the source or destination is blocking.  However, as far as audio
playback chain goes, the sink should ALWAYS be the limiting
factor as decoder sources need to be able to produce data at
least as fast as the audio is being played (otherwise there'll
be audible drops).

Thus, we bias the select(2) into waiting on a targets on if we
splice(2) less than the data we requested.